better backporting code
